requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==4.1.13
django-cors-headers==3.14.0
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.28.0
flake8==7.1.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
pillow==11.2.1
pycodestyle==2.12.1
pyflakes==3.2.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
cryptography